Union of both branches' structural extensions (EntityKind variants, Entity
fields, dispatch arms, tick pipeline), plus the planned INTEGRATE seams:

- town_npc_positions() now returns live §7 NPC centers, making §5.3 step 3
  town spawn suppression real (pinned by town_npcs_suppress_hostile_spawns)
- npcs' stopgap player HP/DebuffSet storage replaced by survival's (§8)
  fields; nurse heal/dialogue/arrival conditions ported to u16 hp +
  Vec<(Debuff, ticks)>
- duplicate set_bed_spawn resolved in favor of sim::survival's
- spawn_message returns Option again (NPCs announce via NpcList); enemy and
  projectile spawn sites adapted
- entity snapshots carry both the NPC anim state byte and enemy hp deltas
- client: one vitals mirror (HUD + nurse cost preview), right-click talk
  composed with combat input (hovered NPC outranks the tile behind it),
  exhaustive ServerMessage match

+++ crates/client/src/ui/dialogue.rs
@@ -131,8 +131,8 @@ impl DialogueUi {
     pub fn frame(
         &mut self,
         inv: &[Option<InvSlot>],
-        own_hp: u32,
-        own_max_hp: u32,
+        own_hp: u16,
+        own_max_hp: u16,
         debuffs: &[ActiveDebuff],
         flags: WorldFlags,
         now: f64,
@@ -189,7 +189,7 @@ impl DialogueUi {
                     ("Heal (healthy)".to_string(), false)
                 } else {
                     let cost = nurse_heal_cost(
-                        hp_restored,
+                        hp_restored as u32,
                         cleared,
                         flags.watcher_defeated,
                         flags.bone_warden_defeated,
